Move integration tests into separate directory, write more unit tests
