+++ oonidata/tests/data/.gitignore
@@ -2,3 +2,5 @@
 /raw_measurements/*
 /measurements/*
 /s3-dump/*
+/postcans/*
+/jsonlgzs/*
